Generate OSGi metadata to make apfloat OSGi compliant
This allows usage of the following modules in an OSGi runtime out of the
box:
- apfloat
- apfloat-aparapi
- apfloat-jscience

Additionally let BND-Tools generate the services file for the apfloat
module together with corresponding OSGi metadata to make service usage
work in OSGi runtimes. In order to make this work add a compile-time
only dependency to 'biz.aQute.bnd:biz.aQute.bnd.annotation'.